+++ wrapper.h
@@ -6,6 +6,8 @@
 #include <open62541/client_highlevel_async.h>
 #include <open62541/client_subscriptions.h>
 #include <open62541/plugin/log_stdout.h>
+// Include with binding of `vsnprintf()` to simplify formatting of log messages.
+#include <stdio.h>
 
 // bindgen does not support non-trivial `#define` used for pointer constant. Use
 // statically defined constant as workaround for now.