+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2024-01-22
+
+[0.2.2]: https://github.com/HMIProject/open62541-sys/compare/v0.2.1...v0.2.2
+
+### Changed
+
+- Add binding for `va_end()`, adjust behavior of `vsnprintf()` to call `va_copy()` internally
+  (HMIProject/open62541#30).
+
 ## [0.2.1] - 2024-01-19
 
 [0.2.1]: https://github.com/HMIProject/open62541-sys/compare/v0.2.0...v0.2.1